Renamed CommandSwerve File moved it to subsystem. - Rishik B.
